+++ setup-lib.py
@@ -71,7 +71,7 @@ ext_modules = [ ]
 if platform_system() == 'Windows':
     ext_modules = [Extension("libopenzwave",
                              sources=["src-lib/libopenzwave/libopenzwave.pyx"],
-                             libraries=['setupapi', 'stdc++'],
+                             libraries=['setupapi', 'msvcrt', 'ws2_32', 'dnsapi'],
                              language="c++",
                              extra_objects=['openzwave/libopenzwave.a'],
                              include_dirs=['openzwave/cpp/src', 'openzwave/cpp/src/value_classes', 'openzwave/cpp/src/platform', 'openzwave/cpp/build/windows', "src-lib/libopenzwave"]