+++ src-api/openzwave/node.py
@@ -27,7 +27,8 @@ from openzwave.object import ZWaveObject
 from openzwave.group import ZWaveGroup
 from openzwave.value import ZWaveValue
 from openzwave.command import ZWaveNodeBasic, ZWaveNodeSwitch
-from openzwave.command import ZWaveNodeSensor, ZWaveNodeSecurity
+from openzwave.command import ZWaveNodeSensor, ZWaveNodeThermostat
+from openzwave.command import ZWaveNodeSecurity
 
 # Set default logging handler to avoid "No handler found" warnings.
 import logging
@@ -43,7 +44,8 @@ logger.addHandler(NullHandler())
 
 class ZWaveNode(ZWaveObject,
                 ZWaveNodeBasic, ZWaveNodeSwitch,
-                ZWaveNodeSensor, ZWaveNodeSecurity):
+                ZWaveNodeSensor, ZWaveNodeThermostat,
+                ZWaveNodeSecurity):
     """
     Represents a single Node within the Z-Wave Network.