+++ src-lib/libopenzwave/libopenzwave.pyx
@@ -72,7 +72,7 @@ logger = logging.getLogger('libopenzwave')
 logger.addHandler(NullHandler())
 
 from pkg_resources import get_distribution, DistributionNotFound
-__version__ = "0.3.0b9"
+__version__ = "0.3.1"
 libopenzwave_location = 'not_installed'
 libopenzwave_file = 'not_installed'
 try:
@@ -185,6 +185,7 @@ PyNotifications = [
     EnumWithDoc('Notification').setDoc("A manager notification report."),
     EnumWithDoc('DriverRemoved').setDoc("The Driver is being removed."),
     EnumWithDoc('ControllerCommand').setDoc("When Controller Commands are executed, Notifications of Success/Failure etc are communicated via this Notification."),
+    EnumWithDoc('NodeReset').setDoc("A node has been reset from OpenZWave's set.  The Device has been reset and thus removed from the NodeList in OZW."),
     ]
 
 PyNotificationCodes = [
@@ -1021,6 +1022,7 @@ sleeping) have been polled, an "AllNodesQueried" notification is sent.
         0x62: 'COMMAND_CLASS_DOOR_LOCK',
         0x63: 'COMMAND_CLASS_USER_CODE',
         0x64: 'COMMAND_CLASS_GARAGE_DOOR',
+        0x66: 'COMMAND_CLASS_BARRIER_OPERATOR',
         0x70: 'COMMAND_CLASS_CONFIGURATION',
         0x71: 'COMMAND_CLASS_ALARM',
         0x72: 'COMMAND_CLASS_MANUFACTURER_SPECIFIC',